javascript warm up

+++ 0x12-javascript-warm_up/13-add.js
@@ -1,10 +1,9 @@
 #!/usr/bin/node
 
 function add (a, b) {
-    const aye = parseFloat(a);
-    const bee = parseFloat(b);
-    return aye + bee;
+    const aye = parseInt(a);
+    const bee = parseInt(b);
+    return (aye + bee);
   };
   
   module.exports = add;
-
